Share benefitsLabels
